# openapi.yaml
openapi: 3.0.0
info:
  title: Ресторан Rest API Спецификация
  version: 0.0.1
servers: 
  - url: https://laravel-rest.ru/api/
    description: Dev server
paths:
...
  /forgot-password:
    post:
      summary: Отправка сообщения на почту с кодом
      tags: 
        - Auth
      operationId: forgotpassword
      description: Отправка сообщения на почту
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPassword"
      responses:
        '200':
          description: Успешная регистрация
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForgotPasswordSuccess"
        default:
          description: Ошибка
          content:
            application/json:
              schema:
               $ref: "#/components/schemas/Error"
...
  /reset-password:
    post:
      summary: Процедура сброса пароля
      tags: 
        - Auth
      operationId: resetPassword
      description: Ввод нового пароля
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPassword"
      responses:
        '200':
          description: Пароль успешно изменен
          content:
            application/json: {}
        '401':
          description: Новый пароль не должен совпадать со старым
          content:
            application/json: {}
        '400':
          description: Ошибка смены пароля
          content:
            application/json: {}
        default:
          description: Пинкод неверный
          content:
            application/json:
              schema:
               $ref: "#/components/schemas/Error" 
...
components:
  schemas:
...
     Error:
        type: object
        required:
          - code
          - message
        properties:
          code:
            type: integer
            example: 500
          message:
            type: string
            example: Что-то пошло не так
...
     ForgotPassword:
      type: object
      required:
        - email
      properties:
          email:
            type: string
            example: dod@300yandex.ru
     ForgotPasswordSuccess:
      type: object
      required:
        - email
      properties:
          message:
            type: string
            example: Письмо успешно отправлено
     ResetPassword:
      type: object
      required:
        - password
        - password_confirmation
        - email
      properties:
          email:
            type: string
            example: ilia@cobaka.ru
          password:
            type: integer
            example: klfsgjlk;sd;j32524
          password_confirmation:
            type: integer
            example: klfsgjlk;sd;j32524
...